requirements.txt
asgiref==3.7.2
attrs==23.2.0
bandit==1.7.6
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
Django~=4.2.9
django-multiselectfield==0.1.12
dparse==0.6.3
fonttools==4.47.2
# Editable install with no version control (foo==1.0)
#-e /home/jason/Documents/MyVintedWizardLocal2/MyVintedWizard/src
gitdb==4.0.11
GitPython==3.1.41
h11==0.14.0
idna==3.6
kiwisolver==1.4.5
markdown-it-py==3.0.0
matplotlib==3.8.2
mdurl==0.1.2
numpy==1.26.3
outcome==1.3.0.post0
packaging==21.3
pandas~=1.5.3
pbr==6.0.0
pillow==10.2.0
plotly==5.18.0
Pygments==2.17.2
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.0
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
safety==2.3.5
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
stevedore==5.1.0
tenacity==8.2.3
trio==0.24.0
trio-websocket==0.11.1
tzdata==2023.4
urllib3==2.1.0
wsproto==1.2.0

setuptools~=69.0.3
loguru~=0.7.2
pika~=1.3.2

schedule~=1.2.1
python-dotenv~=1.0.1
discord~=2.3.2
playwright~=1.42.0